+++ Changelog.md
@@ -7,6 +7,7 @@
 - Add support for DHS/VHS hotkeys for common settings
 - Fix some memory leaks
 - Add `VTank.FixPortalLoops` setting that will attempt to detect and fix infinite portal loops while naving
+- Fix bug where item counts from `/ub count profile` were not properly resetting
 
 ## 0.1.2 (2020-01-02) [UtilityBeltInstaller-0.1.2.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/b1bcdf8401543ba32afd4cf480ff6408/UtilityBeltInstaller-0.1.2.exe)
 - "Fix" `/ub playeroption` on ACE, by sending the entire PlayerSystem packet