+++ Changelog.md
@@ -9,7 +9,8 @@
 - Add `VTank.FixPortalLoops` setting that will attempt to detect and fix infinite portal loops while naving
 - Fix bug where item counts from `/ub count profile` were not properly resetting
 - Add ItemGiver UI
-- Optionally replace vtanks meta expression system with one from UB. Read more here: [TODO_LINK](http://utilitybelty.gitlab.io/) 
+- Optionally replace vtanks meta expression system with one from UB. Read more here: [TODO_LINK](http://utilitybelty.gitlab.io/)
+- Add settings for `AutoVendor.EnableBuying` and `AutoVendor.EnableSelling` to disable specific autovendor functionality
 
 ## 0.1.2 (2020-01-02) [UtilityBeltInstaller-0.1.2.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/b1bcdf8401543ba32afd4cf480ff6408/UtilityBeltInstaller-0.1.2.exe)
 - "Fix" `/ub playeroption` on ACE, by sending the entire PlayerSystem packet