+++ Changelog.md
@@ -1,3 +1,6 @@
+## 0.1.5 (TBD)
+- Insert changes here
+
 ## 0.1.4 [UtilityBeltInstaller-0.1.4.exe](/uploads/5e6eaf2f4dba7d0ee2ab445d110e959b/UtilityBeltInstaller-0.1.4.exe)
 - Fix `/ub count profile <profile>` to actually output results to chat
 - Add `/ub date[utc] [format]` command to echo the current date, with optional format